Refactor so Constants are in their own file
